docs: add ARACA refactor report
