+++ .github/workflows/deploy.yml
@@ -8,6 +8,10 @@ permissions:
   contents: read
   packages: write  # ← Enables write access for GHCR
 
+env:
+  NODE_VERSION: '20'
+  IMAGE_ID: ghcr.io/${{ github.actor }}/effusion-labs:latest
+
 jobs:
   build:
     runs-on: ubuntu-latest
@@ -19,13 +23,24 @@ jobs:
       - name: Setup Node.js
         uses: actions/setup-node@v3
         with:
-          node-version: '20'
+          node-version: ${{ env.NODE_VERSION }}
+
+      - name: Cache npm
+        uses: actions/cache@v4
+        with:
+          path: ~/.npm
+          key: ${{ runner.os }}-npm-${{ env.NODE_VERSION }}-${{ hashFiles('package-lock.json') }}
+          restore-keys: |
+            ${{ runner.os }}-npm-${{ env.NODE_VERSION }}-
 
       - name: Install dependencies
         run: npm ci
 
+      - name: Run tests
+        run: npm test
+
       - name: Build Eleventy static files
-        run: npx @11ty/eleventy
+        run: npm run build
 
       - name: Log in to GitHub Container Registry
         uses: docker/login-action@v3
@@ -35,8 +50,6 @@ jobs:
           password: ${{ secrets.GITHUB_TOKEN }}
 
       - name: Build and push Docker image
-        env:
-          IMAGE_ID: ghcr.io/${{ github.actor }}/effusion-labs:latest
         run: |
           echo "📦 Building and pushing $IMAGE_ID"
           docker build -t "$IMAGE_ID" -f .portainer/Dockerfile .
