+++ ARACA_REPORT_20250731-062043.md
@@ -1,0 +1,24 @@
+# ARACA Refactor Report
+
+**Timestamp:** 2025-07-31T06:20:43Z
+
+## Objectives and Constraints
+- Modularize build components and clean up old configs.
+- Remove legacy references to `main.css` and streamline Docker/GH Actions.
+- Maintain Eleventy, Nunjucks, Tailwind baseline.
+
+## Inventory of Changes
+- `.eleventy.js` now imports shortcodes from new module.
+- `lib/shortcodes.js` added with `specnote` implementation.
+- `.portainer/Dockerfile` simplified to remove obsolete inspection steps.
+- `.github/workflows/deploy.yml` uses cached npm install, runs tests, and centralizes env vars.
+- `package.json` declares Node engine requirement.
+- `.nvmrc` created for version control.
+
+## Validation
+- `npm test` → all 8 tests pass.
+- `npm run build` → Eleventy build succeeds.
+
+## Next Steps
+- Monitor workflow run for reduced build time.
+- Consider adding linter step in CI.
